Clipboard reorganization.
    * Split clipboard handlers into separate functions.
    * The appropriate clipboard is built and assigned at compile time
      (can be changed in the future to allow changes via config).
    * All platforms except for Win32 default to SDL2 if libsdl2 is used.
      System-specific clipboard handlers are avoided where possible.
    * The selected clipboard handler is displayed on the console when
      MegaZeux is started.
