+++ src/editor/clipboard_carbon.h
@@ -18,19 +18,12 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
  */
 
-#ifndef __EDITOR_CLIPBOARD_H
-#define __EDITOR_CLIPBOARD_H
+#ifndef __EDITOR_CLIPBOARD_CARBON_H
+#define __EDITOR_CLIPBOARD_CARBON_H
 
 #include "../compat.h"
 
-__M_BEGIN_DECLS
+void clipboard_set_carbon(char **buffer, int lines, int total_length);
+char *clipboard_get_carbon(void);
 
-void clipboard_init(void);
-
-void copy_buffer_to_clipboard(char **buffer, int lines, int total_length);
-
-char *get_clipboard_buffer(void);
-
-__M_END_DECLS
-
-#endif // __EDITOR_CLIPBOARD_H
+#endif // __EDITOR_CLIPBOARD_CARBON_H
