+++ src/editor/clipboard_sdl2.h
@@ -1,6 +1,5 @@
 /* MegaZeux
  *
- * Copyright (C) 2004 Gilead Kutnick <exophase@adelphia.net>
  * Copyright (C) 2017 Ian Burgmyer <spectere@gmail.com>
  *
  * This program is free software; you can redistribute it and/or
@@ -18,19 +17,12 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
  */
 
-#ifndef __EDITOR_CLIPBOARD_H
-#define __EDITOR_CLIPBOARD_H
+#ifndef __EDITOR_CLIPBOARD_SDL2_H
+#define __EDITOR_CLIPBOARD_SDL2_H
 
 #include "../compat.h"
 
-__M_BEGIN_DECLS
+void clipboard_set_sdl2(char **buffer, int lines, int total_length);
+char *clipboard_get_sdl2(void);
 
-void clipboard_init(void);
-
-void copy_buffer_to_clipboard(char **buffer, int lines, int total_length);
-
-char *get_clipboard_buffer(void);
-
-__M_END_DECLS
-
-#endif // __EDITOR_CLIPBOARD_H
+#endif // __EDITOR_CLIPBOARD_SDL2_H
